# Remove support for skip_rows / num_rows options in the parquet reader.