@vuule 2022-08-09
Please post the impact on the reader benchmarks.