@nvdbaranec 2022-08-11
Just realized the CI failure was because of actual compile errors in the benchmarks.  Fixed.